+++ src/resonaate/tasking/engine/engine_base.py
@@ -1,7 +1,10 @@
 """Abstract :class:`.Tasking` base class defining the tasking engine API."""
+from __future__ import annotations
+
 # Standard Library Imports
 from abc import ABCMeta, abstractmethod
 from logging import getLogger
+from typing import TYPE_CHECKING
 
 # Third Party Imports
 from numpy import zeros
@@ -11,6 +14,13 @@ from ...data.importer_database import ImporterDatabase
 from ..decisions.decision_base import Decision
 from ..rewards.reward_base import Reward
 
+# Type Checking Imports
+if TYPE_CHECKING:
+    # Local Imports
+    from ...data.observation import Observation
+    from ...data.task import Task
+    from ...physics.time.stardate import JulianDate
+
 
 class TaskingEngine(metaclass=ABCMeta):
     """Abstract base class defining common API for tasking engines.
@@ -18,11 +28,19 @@ class TaskingEngine(metaclass=ABCMeta):
     This class provides the framework and behavior for the command & control of a network or agent.
     """
 
-    def __init__(self, engine_id, sensor_ids, target_ids, reward, decision, importer_db_path=None):
+    def __init__(
+        self,
+        engine_id: int,
+        sensor_ids: list[int],
+        target_ids: list[int],
+        reward: Reward,
+        decision: Decision,
+        importer_db_path: str | None = None,
+    ):
         """Initialize a tasking engine object.
 
         Args:
-            engine_id (int): Unique ID for this :class:`.TaskingEngine`
+            engine_id (``int``): Unique ID for this :class:`.TaskingEngine`
             sensor_ids (``list``): list of sensor agent ID numbers
             target_ids (``list``): list of target agent ID numbers
             reward (:class:`.Reward`): callable reward object for determining tasking priority
@@ -75,7 +93,7 @@ class TaskingEngine(metaclass=ABCMeta):
         if importer_db_path:
             self._importer_db = ImporterDatabase.getSharedInterface(db_path=importer_db_path)
 
-    def addTarget(self, target_id):
+    def addTarget(self, target_id: int) -> None:
         """Add a target to this :class:`.TaskingEngine`.
 
         Args:
@@ -84,7 +102,7 @@ class TaskingEngine(metaclass=ABCMeta):
         self.target_indices[target_id] = len(self.target_list)
         self.target_list.append(target_id)
 
-    def removeTarget(self, target_id):
+    def removeTarget(self, target_id: int) -> None:
         """Remove a target from this :class:`.TaskingEngine`.
 
         Args:
@@ -93,7 +111,7 @@ class TaskingEngine(metaclass=ABCMeta):
         del self.target_indices[target_id]
         self.target_list.remove(target_id)
 
-    def addSensor(self, sensor_id):
+    def addSensor(self, sensor_id: int) -> None:
         """Add a sensor to this :class:`.TaskingEngine`.
 
         Args:
@@ -101,7 +119,7 @@ class TaskingEngine(metaclass=ABCMeta):
         """
         self.sensor_list.append(sensor_id)
 
-    def removeSensor(self, sensor_id):
+    def removeSensor(self, sensor_id: int) -> None:
         """Remove a sensor from this :class:`.TaskingEngine`.
 
         Args:
@@ -109,7 +127,7 @@ class TaskingEngine(metaclass=ABCMeta):
         """
         self.sensor_list.remove(sensor_id)
 
-    def saveObservations(self, observations):
+    def saveObservations(self, observations: list[Observation]) -> None:
         """Save set of :class:`.Observation` objects to transient lists.
 
         Args:
@@ -118,14 +136,14 @@ class TaskingEngine(metaclass=ABCMeta):
         self._observations.extend(observations)
         self._saved_observations.extend(observations)
 
-    def getCurrentObservations(self):
+    def getCurrentObservations(self) -> list[Observation]:
         """``list``: Returns current list of observations saved internally & resets transient list."""
         observations = self._saved_observations
         self._saved_observations = []
 
         return observations
 
-    def retaskSensors(self, new_target_nums):
+    def retaskSensors(self, new_target_nums: list[int]) -> None:
         """Update the set of target agents, usually after a target is added/removed.
 
         Args:
@@ -134,7 +152,7 @@ class TaskingEngine(metaclass=ABCMeta):
         self.target_list = new_target_nums
 
     @abstractmethod
-    def assess(self, prior_julian_date, julian_date):
+    def assess(self, prior_julian_date: JulianDate, julian_date: JulianDate) -> None:
         """Perform a set of analysis operations on the current simulation state.
 
         Must be overridden by implemented classes.
@@ -146,7 +164,7 @@ class TaskingEngine(metaclass=ABCMeta):
         raise NotImplementedError
 
     @abstractmethod
-    def generateTasking(self):
+    def generateTasking(self) -> None:
         """Create tasking solution based on the current simulation state.
 
         Must be overridden by implemented classes.
@@ -154,7 +172,7 @@ class TaskingEngine(metaclass=ABCMeta):
         raise NotImplementedError
 
     @abstractmethod
-    def getCurrentTasking(self, julian_date):
+    def getCurrentTasking(self, julian_date: JulianDate) -> Task:
         """Return current tasking solution.
 
         Must be overridden by implemented classes.
@@ -168,31 +186,31 @@ class TaskingEngine(metaclass=ABCMeta):
         raise NotImplementedError
 
     @property
-    def reward(self):
+    def reward(self) -> Reward:
         """:class:`.Reward`: Returns the tasking engine's reward function."""
         return self._reward
 
     @property
-    def decision(self):
+    def decision(self) -> Decision:
         """:class:`.Decision`: Returns the tasking engine's decision function."""
         return self._decision
 
     @property
-    def unique_id(self):
+    def unique_id(self) -> int:
         """int: Unique identifier for this :class:`.TaskingEngine`."""
         return self._unique_id
 
     @property
-    def num_targets(self):
+    def num_targets(self) -> int:
         """``int``: Returns the number of targets."""
         return len(self.target_list)
 
     @property
-    def num_sensors(self):
+    def num_sensors(self) -> int:
         """``int``: Returns the number of sensors."""
         return len(self.sensor_list)
 
     @property
-    def observations(self):
+    def observations(self) -> list[Observation]:
         """``list``: Returns the :class:`.Observation` objects for the previous timestep."""
         return self._observations