+++ src/resonaate/tasking/rewards/reward_base.py
@@ -1,22 +1,33 @@
 """Abstract :class:`.Reward` base class defining the reward API."""
+from __future__ import annotations
+
 # Standard Library Imports
 from abc import ABCMeta, abstractmethod
+from typing import TYPE_CHECKING
 
 # Local Imports
 from ..metrics.metric_base import Metric
+
+if TYPE_CHECKING:
+    # Third Party Imports
+    from numpy import ndarray
+
+    # Local Imports
+    from ...agents.estimate_agent import EstimateAgent
+    from ...agents.sensing_agent import SensingAgent
 
 
 class Reward(metaclass=ABCMeta):
     """Abstract base class to encapsulate behavior of general reward methods."""
 
-    REGISTRY = {}
-    """Global reward object registry."""
+    REGISTRY: dict[str, "Reward"] = {}
+    """``dict``: Global reward object registry."""
 
-    def __init__(self, metrics):
+    def __init__(self, metrics: list[Metric]):
         """Construct a reward object with a set of metrics.
 
         Args:
-            metrics (list): metrics to include in the reward calculation
+            metrics (``list``): :class:`.Metric` objects to include in the reward calculation
 
         Raises:
             TypeError: raised if invalid :class:`.Metric` objects are passed
@@ -28,11 +39,11 @@ class Reward(metaclass=ABCMeta):
         self._metrics = metrics
 
     @classmethod
-    def register(cls, name, reward):
+    def register(cls, name: str, reward: "Reward") -> None:
         """Register an implemented reward class in the global registry.
 
         Args:
-            name (str): name to store as the key in the registry
+            name (``str``): name to store as the key in the registry
             reward (:class:`.Reward`): reward object to register
 
         Raises:
@@ -43,63 +54,61 @@ class Reward(metaclass=ABCMeta):
         cls.REGISTRY[name] = reward
 
     @property
-    def is_registered(self):
+    def is_registered(self) -> bool:
         """bool: return if an implemented reward class is registered."""
         return self.__class__.__name__ in self.REGISTRY
 
-    def calculateMetrics(self, target_agents, target_id, sensor_agents, sensor_id, **kwargs):
+    def calculateMetrics(
+        self, estimate_agent: EstimateAgent, sensor_agent: SensingAgent, **kwargs
+    ) -> dict[Metric, float]:
         """Calculate each metric and saves to a dictionary.
 
         Args:
-            target_agents (dict): current target agents set
-            target_id (int): unique id of target corresponding to the metric
-            sensor_agents (dict): current sensor agents set
-            sensor_id (int): unique id of sensor corresponding to the metric
+            estimate_agent (:class:`.EstimateAgent`): estimate agent for which this metric is being calculated
+            sensor_agent (:class:`.SensorAgent`): sensor agent for which this metric is being calculated
 
         Returns:
-            dict: calculated metrics for the current simulation state
+            (``dict``): calculated metrics for the current simulation state
         """
         metric_solutions = {}
         for metric in self.metrics:
-            metric_solutions[metric] = metric(
-                target_agents, target_id, sensor_agents, sensor_id, **kwargs
-            )
+            metric_solutions[metric] = metric(estimate_agent, sensor_agent, **kwargs)
 
         return metric_solutions
 
     @abstractmethod
-    def _calculateReward(self, target_agents, target_id, sensor_agents, sensor_id, **kwargs):
+    def _calculateReward(
+        self, estimate_agent: EstimateAgent, sensor_agent: SensingAgent, **kwargs
+    ) -> ndarray:
         """Abstract function for calculating rewards based on the simulation state.
 
         Note:
             Must be overridden by implementors.
 
         Args:
-            target_agents (dict): current target agents set
-            target_id (int): unique id of target corresponding to the metric
-            sensor_agents (dict): current sensor agents set
-            sensor_id (int): unique id of sensor corresponding to the metric
+            estimate_agent (:class:`.EstimateAgent`): estimate agent for which this metric is being calculated
+            sensor_agent (:class:`.SensorAgent`): sensor agent for which this metric is being calculated
 
         Returns:
-            ``numpy.ndarray``: calculated reward
+            ``ndarray``: calculated reward
         """
         raise NotImplementedError
 
-    def __call__(self, target_agents, target_id, sensor_agents, sensor_id, **kwargs):
+    def __call__(
+        self, estimate_agent: EstimateAgent, sensor_agent: SensingAgent, **kwargs
+    ) -> ndarray:
         """Call operator '()' for reward objects.
 
         Args:
-            target_agents (dict): current target agents set
-            target_id (int): unique id of target corresponding to the metric
-            sensor_agents (dict): current sensor agents set
-            sensor_id (int): unique id of sensor corresponding to the metric
+            estimate_agent (:class:`.EstimateAgent`): estimate agent for which this metric is being calculated
+            sensor_agent (:class:`.SensorAgent`): sensor agent for which this metric is being calculated
 
         Returns:
-            ``numpy.ndarray``: calculated reward
+            ``ndarray``: calculated reward
         """
-        return self._calculateReward(target_agents, target_id, sensor_agents, sensor_id, **kwargs)
+        return self._calculateReward(estimate_agent, sensor_agent, **kwargs)
 
     @property
-    def metrics(self):
-        """list: metric objects used to calculate the reward."""
+    def metrics(self) -> list[Metric]:
+        """``list``: metric objects used to calculate the reward."""
         return self._metrics