+++ CHANGELOG.md
@@ -29,7 +29,7 @@ ______________________________________________________________________
 # \[Unreleased\]
 
 - Added
-  
+
   - `getStandardDeviation` and `getConfidenceRegion` functions in `resonaate.physics.statistics`
   - `fov` input variable to `sensorFactory()`
   - EOPs from 06-10-2021 to 17-03-2022
@@ -51,13 +51,17 @@ ______________________________________________________________________
   - `AsyncExecuteTasking()` now calls `collectObservations` instead of `makeNoisyObservation`
   - chi-square test functions now accept `ndarray` as inputs
   - `_getSolarRadiationPressureAcceleration()` is now a class function of `SpecialPerturbations`
+  - Metric calculations only take in 1 estimate agent and sensor agent
 
 - Deprecated
+
 - Removed
 
   - `DEFAULT_VIZ_X_SECTION` and all references to it
 
 - Fixed
+
+  - FOV tests now reliably pass, properly call `updateReductionParameters`
 
 # \[1.4.0 - 2022-06-06\]